Update API scheme to layer 143.

+++ Telegram/SourceFiles/api/api_peer_photo.cpp
@@ -75,7 +75,7 @@ SendMediaReady PreparePeerPhoto(
 		MTP_int(dcId));
 
 	QString file, filename;
-	int32 filesize = 0;
+	int64 filesize = 0;
 	QByteArray data;
 
 	return SendMediaReady(
