+++ Telegram/SourceFiles/media/streaming/media_streaming_loader_local.h
@@ -21,10 +21,10 @@ public:
 	LoaderLocal(std::unique_ptr<QIODevice> device);
 
 	[[nodiscard]] Storage::Cache::Key baseCacheKey() const override;
-	[[nodiscard]] int size() const override;
+	[[nodiscard]] int64 size() const override;
 
-	void load(int offset) override;
-	void cancel(int offset) override;
+	void load(int64 offset) override;
+	void cancel(int64 offset) override;
 	void resetPriorities() override;
 	void setPriority(int priority) override;
 	void stop() override;
@@ -42,7 +42,7 @@ private:
 	void fail();
 
 	const std::unique_ptr<QIODevice> _device;
-	const int _size = 0;
+	const int64 _size = 0;
 	rpl::event_stream<LoadedPart> _parts;
 
 };
